+++ src/gdalraster.cpp
@@ -38,6 +38,22 @@ GDALRaster::GDALRaster(std::string filename, bool read_only) :
 	hDataset = GDALOpenShared(fname.c_str(), eAccess);
 	if (hDataset == NULL)
 		Rcpp::stop("Open raster failed.");
+	
+	// warn for now if 64-bit integer
+	bool has_int64 = false;
+	for (int b = 1; b <= this->getRasterCount(); ++b) {
+		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, b);
+		GDALDataType eDT = GDALGetRasterDataType(hBand);
+		if (GDALDataTypeIsInteger(eDT) && (GDALGetDataTypeSizeBits(eDT) == 64))
+			has_int64 = true;
+	}
+	if (has_int64) {
+		Rcpp::Rcout << "Int64/UInt64 raster data types not fully supported.\n";
+		Rcpp::Rcout << "Loss of precision will occur for values > 2^53.\n";
+		std::string msg = 
+			"Int64/UInt64 raster data are currently handled as double.";
+		Rcpp::warning(msg);
+	}
 }
 
 std::string GDALRaster::getFilename() const {