+++ NEWS.md
@@ -1,5 +1,9 @@
 # gdalraster 1.4.1 (dev)
 
+* `GDALRaster$new()`: add warning on open if raster has a 64-bit integer data type (handled as double for now) (2023-09-23)
+
+* fix: `warp()` would segfault if proj.db could not be found (return value of `GDALWarpAppOptionsNew()` was not being checked) (2023-09-19)
+
 * `plot_raster()`: a legend can now be used with a color table for continuous data. Default value of the `legend` argument has been changed to `FALSE` (2023-09-17)
 
 * add `CmbTable$asMatrix()`: returns the table of combinations as a numeric matrix with named columns (alternative to `$asDataFrame()`) (2023-09-15)